[Fix/#158] make components
HomeView뿐만 아니라 MainView에서 LTCICard가 사용되므로 관련된 뷰들을 컴포넌트화시켰다
